Renaming rate measurer and request load with underscore to make them private
