+++ benchmark/scenarios/_rate_measurer.py
