+++ benchmark/scenarios/read_request_load.py
@@ -8,9 +8,9 @@ from zope.interface import implementer
 from twisted.internet.defer import Deferred
 from .._interfaces import IScenario, IRequestGenerator, IScenarioSetup
 
-from rate_measurer import DEFAULT_SAMPLE_SIZE
+from ._rate_measurer import DEFAULT_SAMPLE_SIZE
 
-from request_load import (
+from ._request_load import (
     RequestLoadScenario, RequestOverload, RequestRateTooLow,
     RequestRateNotReached
 )
